+++ backend/src/main/java/com/codecool/trv/service/UserService.java
@@ -4,6 +4,7 @@ import com.codecool.trv.dto.user.NewUserRequest;
 import com.codecool.trv.dto.user.UpdateUserRequest;
 import com.codecool.trv.dto.user.UserResponse;
 import com.codecool.trv.exception.ResourceNotFoundException;
+import com.codecool.trv.mapper.UserMapper;
 import com.codecool.trv.model.Note;
 import com.codecool.trv.model.User;
 import com.codecool.trv.repository.UserRepository;
@@ -25,33 +26,36 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public List<UserResponse> findAllUsers() {
-        List<User> users = userRepository.findAll();
-        return users.stream().map(user -> new UserResponse(user.getId(), user.getUsername())).toList();
-    }
-
-    public UserResponse findUserResponseById(Long id) {
-        User user = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
-        return new UserResponse(user.getId(), user.getUsername());
-    }
-
     User findUserById(Long id) {
         return userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
     }
 
+    Set<User> findUsersByIds(Set<Long> userIds) {
+        return new HashSet<>(userRepository.findAllById(userIds));
+    }
+
+    public List<UserResponse> findAllUsers() {
+        List<User> users = userRepository.findAll();
+        return users.stream().map(UserMapper::mapToUserResponse).toList();
+    }
+
+    public UserResponse findUserResponseById(Long id) {
+        User user = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
+        return UserMapper.mapToUserResponse(user);
+    }
+
+    public UserResponse findUserResponseByName(String name) {
+        User user = userRepository.findUserByUsername(name);
+        if(user == null) {
+            throw new ResourceNotFoundException("User not found");
+        }
+        return UserMapper.mapToUserResponse(user);
+    }
+
     public UserResponse addUser(NewUserRequest newUserRequest) {
         //TODO request validation!!! - unique username, unique email
-        User userToSave = User
-                .builder()
-                .username(newUserRequest.username())
-                .firstName(newUserRequest.firstName())
-                .lastName(newUserRequest.lastName())
-                .email(newUserRequest.email())
-                .password(newUserRequest.password())
-                .build();
-
-        User user = userRepository.save(userToSave);
-        return new UserResponse(user.getId(), user.getUsername());
+        User user = userRepository.save(UserMapper.mapToUser(newUserRequest));
+        return UserMapper.mapToUserResponse(user);
     }
 
     public UserResponse updateUserById(Long id, UpdateUserRequest updateUserRequest) {
@@ -65,28 +69,15 @@ public class UserService {
         userToUpdate.setEmail(updateUserRequest.email());
         userToUpdate.setPassword(updateUserRequest.password());
 
-        userRepository.save(userToUpdate);
-        return new UserResponse(userToUpdate.getId(), userToUpdate.getUsername());
+        User user = userRepository.save(userToUpdate);
+        return UserMapper.mapToUserResponse(user);
     }
 
-    public void deleteUserById(Long id) throws EmptyResultDataAccessException {
-        //TODO maybe not delete the whole entity,
-        // let's create a field boolean active, which marks whether the user is active or not
-        // TODO do we want to delete all journals and notes owned by the user also?
-        // - or just the ones of which he is the owner and there are no contributors.
-        // and the notes which where left as contributor will be marked as anonymous...
-        // cascadeType check!!!
+    public void deleteUserById(Long id) {
+        if(!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User not found");
+        }
         userRepository.deleteById(id);
-    }
-
-    //FIXME: This is only for testing purposes
-    public void deleteAllUsers() {
-        userRepository.deleteAll();
-    }
-
-
-    public Set<User> findUsersByIds(Set<Long> userIds) {
-        return new HashSet<>(userRepository.findAllById(userIds));
     }
 
 }
